requirements.txt
discord.py==2.3.2
yt-dlp==2023.12.30
spotipy==2.23.0
python-dotenv==1.0.0
aiosqlite==0.19.0
aiohttp==3.9.1
deezer-py==1.3.14
